refactor(vue): split markdown node rendering
