+++ test/vue/markdown-renderer.test.ts
@@ -4,7 +4,7 @@ import type { MarkdownNode as Node } from '../../packages/vue/src/types'
 import { flushPromises, mount } from '@vue/test-utils'
 import { describe, expect, it, vi } from 'vitest'
 import { defineComponent, h, nextTick, shallowRef } from 'vue'
-import ComarkNodeList from '../../packages/vue/src/components/renderers/node-list'
+import MarkdownNodes from '../../packages/vue/src/components/renderers/markdown-nodes'
 import { useContext } from '../../packages/vue/src/composables'
 
 function mountNodes(nodes: ShallowRef<Node[]>, components = {}) {
@@ -15,7 +15,7 @@ function mountNodes(nodes: ShallowRef<Node[]>, components = {}) {
         enableCaret: false,
         mode: 'static',
       })
-      return () => h(ComarkNodeList, {
+      return () => h(MarkdownNodes, {
         components,
         nodes: nodes.value,
       })
@@ -25,7 +25,7 @@ function mountNodes(nodes: ShallowRef<Node[]>, components = {}) {
   return mount(Root)
 }
 
-describe('comark renderer', () => {
+describe('markdown renderer', () => {
   it('renders semantic nodes and bound attributes directly', async () => {
     const nodes = shallowRef<Node[]>([
       ['h2', { id: 'title' }, ['strong', {}, 'Title']],
